Fixed movement lag

+++ lib/rendering/Kamera.h
@@ -8,7 +8,7 @@ typedef class Renderable Renderable;
 
 class Kamera {
 private:
-	int y, x;
+	float y, x;
 	int w, h;
 	std::shared_ptr<Renderable> following;
 
@@ -18,8 +18,8 @@ public:
 
 	void update();
 
-	int getX();
-	int getY();
+	float getX();
+	float getY();
 };
 
 #endif // !KAMERA
